Move CheckListItem to common-ui
