+++ common/common-ui/src/main/res/drawable/ic_check_green_round_16.xml
