+++ common/common-ui/src/main/res/drawable/ic_check_grey_round_16.xml
