+++ common/common-ui/src/main/res/drawable/ic_exclamation_yellow_16.xml
