+++ common/common-ui/src/main/res/layout/view_check_list_item.xml
